Re-hash legacy plain-text passwords and stop attempt ids from wrapping
Two defects found while reading the Java backend.

Plain-text passwords survived migration permanently. The legacy schema
stored them as VARCHAR(15), which can hold nothing else, and V2 only
widened the column to fit hashes without converting any value.
LegacyPasswordEncoder.matches accepts an unprefixed value by comparing it
to the raw password, upgradeEncoding was not overridden, and nothing on the
login path re-encoded. Only registration and an explicit password change
ever produced a hash, so a migrated account kept a readable password no
matter how often it signed in — while the README claimed passwords were
stored as salted PBKDF2.

Overriding upgradeEncoding alone would have been dead code:
DaoAuthenticationProvider consults the encoder either way but only rewrites
the password when a UserDetailsPasswordService bean exists. ApiUserDetails
Service now implements it, so a successful login replaces the stored value.
The test fixture is itself in the migrated state — every user has a
plain-text password, which is how the suite has been authenticating — so
the new test asserts the value is readable before the login and a PBKDF2
hash after, that the account still works, and that a second login does not
re-encode what is already current.

Attempt ids narrowed from long to int. The controller bound
@PathVariable @Positive long and the service cast it with (int), so
/api/v1/attempts/4294967297 passed validation and resolved to attempt 1.
The query filters by owner, so nothing leaked between accounts, but a
request addressed to a non-existent attempt read — and on the complete
endpoint irreversibly finished — a different one.

Attempt ids are Integer in the schema, so the long was never meaningful.
The controller now binds int, which rejects an out-of-range value at
binding exactly as /quizzes/{quizId} already did, and AttemptResponse and
AttemptCompletionResponse carry int as well. That last part is what
removed the cast rather than relocating it: the records returning long are
why the concurrency test held the id in a long to begin with.

+++ api/src/main/java/ua/nure/latysh/quizzes/api/attempt/AttemptController.java
@@ -40,7 +40,7 @@ public class AttemptController {
     @GetMapping("/attempts/{attemptId}")
     @Operation(summary = "Read an owned quiz attempt without exposing correct answers")
     public AttemptResponse get(
-            @PathVariable @Positive long attemptId,
+            @PathVariable @Positive int attemptId,
             Authentication authentication) {
         return attemptService.findOwned(attemptId, authentication.getName());
     }
@@ -48,7 +48,7 @@ public class AttemptController {
     @PostMapping("/attempts/{attemptId}/complete")
     @Operation(summary = "Submit selected answers and complete an owned attempt once")
     public AttemptCompletionResponse complete(
-            @PathVariable @Positive long attemptId,
+            @PathVariable @Positive int attemptId,
             @Valid @RequestBody CompleteAttemptRequest request,
             Authentication authentication) {
         return attemptService.complete(attemptId, authentication.getName(), request.answerIds());
